+++ test/integration/libvirt/cases/hostname/1.drift.sh
@@ -1,0 +1,4 @@
+assert_remote "initial static hostname differs from desired hostname" \
+  "test \"\$(hostnamectl --static)\" != 'dbf-hostname'"
+assert_remote "initial static hostname is the cloud-init default" \
+  "test \"\$(hostnamectl --static)\" = 'debianform-ci'"